+++ atomik/src/androidMain/kotlin/me/kevinschildhorn/atomik/color/SharedColor.kt
@@ -1,0 +1,12 @@
+package me.kevinschildhorn.atomik.color
+
+import androidx.compose.ui.graphics.Color
+
+actual class SharedColor actual constructor(hex: Long) {
+    var androidColor: Color
+        private set
+
+    init {
+        this.androidColor = Color(hex)
+    }
+}
