+++ atomik/src/androidMain/kotlin/me/kevinschildhorn/atomik/font/+SharedWeight.kt
@@ -1,0 +1,9 @@
+package me.kevinschildhorn.atomik.font
+
+import androidx.compose.ui.text.font.FontWeight
+
+val SharedWeight.fontWeight: FontWeight
+    get() = when (this) {
+        SharedWeight.BOLD -> FontWeight.Bold
+        else -> FontWeight.Normal
+    }
