+++ atomik/src/commonMain/kotlin/me/kevinschildhorn/atomik/atoms/Atom.kt
@@ -1,0 +1,7 @@
+package me.kevinschildhorn.atomik.atoms
+
+import me.kevinschildhorn.atomik.atoms.interfaces.AtomType
+
+abstract class Atom {
+    abstract val type: AtomType
+}
