+++ atomik/src/commonMain/kotlin/me/kevinschildhorn/atomik/atoms/interfaces/AtomType.kt
@@ -1,0 +1,8 @@
+package me.kevinschildhorn.atomik.atoms.interfaces
+
+enum class AtomType {
+    BUTTON,
+    TEXT,
+    IMAGE,
+    TEXTFIELD
+}
