+++ atomik/src/commonMain/kotlin/me/kevinschildhorn/atomik/atoms/interfaces/ColorAtom.kt
@@ -1,0 +1,7 @@
+package me.kevinschildhorn.atomik.atoms.interfaces
+
+import me.kevinschildhorn.atomik.color.SharedColor
+
+interface ColorAtom {
+    val color: SharedColor
+}
