+++ atomik/src/commonMain/kotlin/me/kevinschildhorn/atomik/atoms/interfaces/EnablableAtom.kt
@@ -1,0 +1,7 @@
+package me.kevinschildhorn.atomik.atoms.interfaces
+
+import me.kevinschildhorn.atomik.color.SharedEnabledColor
+
+interface EnablableAtom {
+    val enabledColor: SharedEnabledColor
+}
