+++ atomik/src/commonMain/kotlin/me/kevinschildhorn/atomik/atoms/interfaces/ResourceAtom.kt
@@ -1,0 +1,5 @@
+package me.kevinschildhorn.atomik.atoms.interfaces
+
+interface ResourceAtom<T> {
+    val image: T
+}
