+++ atomik/src/commonMain/kotlin/me/kevinschildhorn/atomik/atoms/interfaces/TextAtom.kt
@@ -1,0 +1,9 @@
+package me.kevinschildhorn.atomik.atoms.interfaces
+
+import me.kevinschildhorn.atomik.color.SharedColor
+import me.kevinschildhorn.atomik.font.SharedTypography
+
+interface TextAtom {
+    val textColor: SharedColor
+    val font: SharedTypography
+}
