+++ atomik/src/commonMain/kotlin/me/kevinschildhorn/atomik/color/SharedColor.kt
@@ -1,0 +1,3 @@
+package me.kevinschildhorn.atomik.color
+
+expect class SharedColor(hex: Long)
