+++ atomik/src/commonMain/kotlin/me/kevinschildhorn/atomik/color/SharedColorSet.kt
@@ -1,0 +1,15 @@
+package me.kevinschildhorn.atomik.color
+
+interface SharedColorSet {
+    val primary: SharedColor
+    val secondary: SharedColor
+    val background: SharedColor
+    val surface: SharedColor
+    val error: SharedColor
+
+    val primaryText: SharedColor
+    val secondaryText: SharedColor
+    val backgroundText: SharedColor
+    val surfaceText: SharedColor
+    val errorText: SharedColor
+}
