+++ atomik/src/commonMain/kotlin/me/kevinschildhorn/atomik/color/SharedEnabledColor.kt
@@ -1,0 +1,6 @@
+package me.kevinschildhorn.atomik.color
+
+data class SharedEnabledColor(
+    val color: SharedColor,
+    val disabledColor: SharedColor,
+)
