+++ atomik/src/commonMain/kotlin/me/kevinschildhorn/atomik/font/SharedTypography.kt
@@ -1,0 +1,3 @@
+package me.kevinschildhorn.atomik.font
+
+data class SharedTypography(val fontName: String, val weight: SharedWeight, val size: Int)
