+++ atomik/src/commonMain/kotlin/me/kevinschildhorn/atomik/font/SharedWeight.kt
@@ -1,0 +1,7 @@
+package me.kevinschildhorn.atomik.font
+
+enum class SharedWeight {
+    NORMAL,
+    BOLD,
+    ITALIC
+}
