+++ atomik/src/commonMain/kotlin/me/kevinschildhorn/atomik/molecules/BaseMolecule.kt
@@ -1,0 +1,7 @@
+package me.kevinschildhorn.atomik.molecules
+
+import me.kevinschildhorn.atomik.atoms.Atom
+
+abstract class BaseMolecule {
+    abstract val atoms: List<Atom>
+}
