--- common/src/androidMain/kotlin/me/kevinschildhorn/common/design/atomicLib/font/+SharedTypography.kt
@@ -1,13 +1,0 @@
-package me.kevinschildhorn.common.design.atomicLib.font
-
-import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.font.FontFamily
-import androidx.compose.ui.unit.sp
-
-
-fun SharedTypography.asComposeTextStyle(fontFamily: FontFamily) =
-    TextStyle(
-        fontFamily = fontFamily,
-        fontWeight = this.weight.fontWeight,
-        fontSize = this.size.sp,
-    )
