--- common/src/commonMain/kotlin/me/kevinschildhorn/common/design/atomicLib/atoms/interfaces/ColorAtom.kt
@@ -1,7 +1,0 @@
-package me.kevinschildhorn.common.design.atomicLib.atoms.interfaces
-
-import me.kevinschildhorn.common.design.atomicLib.color.SharedColor
-
-interface ColorAtom {
-    val color: SharedColor
-}
