--- common/src/commonMain/kotlin/me/kevinschildhorn/common/design/atomicLib/atoms/interfaces/EnablableAtom.kt
@@ -1,7 +1,0 @@
-package me.kevinschildhorn.common.design.atomicLib.atoms.interfaces
-
-import me.kevinschildhorn.common.design.atomicLib.color.SharedEnabledColor
-
-interface EnablableAtom {
-    val enabledColor: SharedEnabledColor
-}
