--- common/src/commonMain/kotlin/me/kevinschildhorn/common/design/atomicLib/atoms/interfaces/ResourceAtom.kt
@@ -1,5 +1,0 @@
-package me.kevinschildhorn.common.design.atomicLib.atoms.interfaces
-
-interface ResourceAtom<T> {
-    val image: T
-}
