--- common/src/commonMain/kotlin/me/kevinschildhorn/common/design/atomicLib/color/SharedColor.kt
@@ -1,3 +1,0 @@
-package me.kevinschildhorn.common.design.atomicLib.color
-
-expect class SharedColor(hex: Long)
