--- common/src/commonMain/kotlin/me/kevinschildhorn/common/design/atomicLib/color/SharedEnabledColor.kt
@@ -1,6 +1,0 @@
-package me.kevinschildhorn.common.design.atomicLib.color
-
-data class SharedEnabledColor(
-    val color: SharedColor,
-    val disabledColor: SharedColor,
-)
