--- common/src/commonMain/kotlin/me/kevinschildhorn/common/design/atomicLib/font/SharedTypography.kt
@@ -1,3 +1,0 @@
-package me.kevinschildhorn.common.design.atomicLib.font
-
-data class SharedTypography(val fontName: String, val weight: SharedWeight, val size: Int)
