--- common/src/commonMain/kotlin/me/kevinschildhorn/common/design/atomicLib/font/SharedWeight.kt
@@ -1,7 +1,0 @@
-package me.kevinschildhorn.common.design.atomicLib.font
-
-enum class SharedWeight {
-    NORMAL,
-    BOLD,
-    ITALIC
-}
