--- common/src/commonMain/kotlin/me/kevinschildhorn/common/design/atomicLib/molecules/BaseMolecule.kt
@@ -1,7 +1,0 @@
-package me.kevinschildhorn.common.design.atomicLib.molecules
-
-import me.kevinschildhorn.common.design.atomicLib.atoms.Atom
-
-abstract class BaseMolecule {
-    abstract val atoms: List<Atom>
-}
