+++ common/src/commonMain/kotlin/me/kevinschildhorn/common/design/theme/UIDesign.kt
@@ -1,6 +1,6 @@
 package me.kevinschildhorn.common.design.theme
 
-import me.kevinschildhorn.common.design.atomicLib.color.SharedColorSet
+import me.kevinschildhorn.atomik.color.SharedColorSet
 
 object UIDesign {
     fun color(light: Boolean = true): SharedColorSet =
