+++ ios/FotoPresenter/uilayer/swiftui/screens/LoginScreen.swift
@@ -18,6 +18,7 @@ struct LoginScreen: View {
 
     var body: some View {
         Text("Hello")
+        Font.title
     }
 }
